# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Amazon",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Get all roles",
                "description": "Returns all roles",
                "operationId": "bf78d43aa8b5c576b78b08b73c404cb6",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/rol/{id}": {
            "put": {
                "tags": [
                    "Roles"
                ],
                "summary": "Update existing rol",
                "description": "Returns updated rol data",
                "operationId": "b807d340545f43c929f1422bdc8fc690",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "type": "string",
                                        "format": "text",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "Delete existing rol",
                "description": "Deletes a record and returns no content",
                "operationId": "8125fad829dc94c51a6530a4af4f9bee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/miSubscription": {
            "get": {
                "tags": [
                    "Subscripciones"
                ],
                "summary": "Get your subscription",
                "description": "Returns  your subscription",
                "operationId": "3dc59e42eb83d0e9579b31afa9609772",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/subscription/{id}": {
            "put": {
                "tags": [
                    "Subscripciones"
                ],
                "summary": "Update existing subscripcion",
                "description": "Returns updated subscripcion data",
                "operationId": "c163284b8b7fcdc8fb44a789086ca17f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "type": "string",
                                        "format": "text",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/susbcription/{id}": {
            "delete": {
                "tags": [
                    "Subscripciones"
                ],
                "summary": "Delete existing subscripcion",
                "description": "Deletes a record and returns no content",
                "operationId": "29dc3e7ac14e3be8573ee7e727961bce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/subscription/": {
            "post": {
                "tags": [
                    "Subscripciones"
                ],
                "description": "Add subscripcion",
                "operationId": "91e86e9608aba07f4c01ff1d4b278ee5",
                "requestBody": {
                    "description": "Pass pedido info",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plan",
                                    "fechainicio",
                                    "fechacaducidad"
                                ],
                                "properties": {
                                    "direccion": {
                                        "type": "string",
                                        "format": "text",
                                        "example": ""
                                    },
                                    "fechainicio": {
                                        "type": "string",
                                        "format": "text",
                                        "example": ""
                                    },
                                    "fechacaducidad": {
                                        "type": "string",
                                        "format": "text",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong data introduced"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}